backend/requirements.txt
fastapi==0.63.0
typing
twilio
uvicorn[standard]
pydantic==1.7.3
python-dotenv==0.15.0
python-multipart
requests
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.12.3
httptools==0.1.1
httpx==0.16.1

# google client
google-cloud-storage
